% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run}
\alias{run}
\title{The core function to interact with DateLife via the web or API}
\usage{
run(input, format = "mrca", partialfield = "liberal", usetnrsfield = "no",
  approximatematchfield = "no")
}
\arguments{
\item{input}{The string to pass to R}

\item{format}{The return format desired}

\item{partialfield}{To be liberal or conservative with returning trees that only match some of the taxa}

\item{usetnrsfield}{Yes or no to use taxonomic name resolution}

\item{approximatematchfield}{Yes or no to use approximate name matching}
}
\value{
A message in the appropriate format
}
\description{
The core function to interact with DateLife via the web or API
}

